LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-anchor-nesting.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: The string did not match the expected pattern.
Popover 1 Inside popover 2  This is popover 1 Popover 2

Harness Error (FAIL), message = Unhandled rejection: The string did not match the expected pattern.
